requirements.txt
-f /usr/share/pip-wheels
beautifulsoup4==4.7.1
Django==2.1.5
django-appconf==1.0.3
django-htmlmin==0.10.0
django-imagekit==4.0.2
html5lib==1.0.1
mysqlclient==1.4.1
pilkit==2.0
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
six==1.12.0
soupsieve==1.8
webencodings==0.5.1
